+++ Scheduler/ViewModel/MainPageViewModel.cs
@@ -9,83 +9,14 @@ using Xamarin.Forms;
 
 namespace Scheduler.ViewModel
 {
-    public class MainPageViewModel : INotifyPropertyChanged
+    public class MainPageViewModel 
     {
-        public event PropertyChangedEventHandler PropertyChanged;
+        
 
-        private IScheduleService<ScheduleRecord> _schedulerService;
-        private ObservableCollection<ScheduleRecord> _listOfItems;
-
-
-        public INavigation Navigation { get; set; }
-
-        public ObservableCollection<ScheduleRecord> ListOfItems
+        public MainPageViewModel()
         {
-            get
-            {
-                return _listOfItems;
-            }
-            set
-            {
-                _listOfItems = value;
-                PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(nameof(ListOfItems)));
-            }
         }
 
-        public ScheduleRecord SelectedItem { get; set; }
-        public Command DeleteCommand { get; set; }
-        public Command AddRecordCommand { get; set; }
-        public Command EditRecordCommand { get; set; }
-
-        public MainPageViewModel(INavigation navigation)
-        {
-            DeleteCommand = new Command(() => OnDeleteTapped());
-            AddRecordCommand = new Command(OnAddRecordTapped);
-            EditRecordCommand = new Command(OnEditRecordTapped);
-            _schedulerService = new SchedulerService<ScheduleRecord>();
-
-            Navigation = navigation;
-
-            InitializeList();
-        }
-
-        public void InitializeList()
-        {
-            ListOfItems = new ObservableCollection<ScheduleRecord>(_schedulerService.GetAll());
-
-        }
-
-        private async Task OnDeleteTapped()
-        {
-            ConfirmConfig config = new ConfirmConfig()
-            {
-                Message = "Delete the record?",
-                OkText = "Delete",
-                CancelText = "Cancel"
-            };
-
-            var res = await GetConfirmResult(config);
-
-            if (res)
-            {
-                _schedulerService.DeleteObject(SelectedItem);
-                ListOfItems.Remove(SelectedItem);
-            }
-        }
-
-        private void OnAddRecordTapped()
-        {
-            Navigation.PushModalAsync(new AddRecordPage(this));
-        }
-
-        private void OnEditRecordTapped()
-        {
-            Navigation.PushModalAsync(new EditRecordPage(this, SelectedItem));
-        }
-
-        private async Task<bool> GetConfirmResult(ConfirmConfig config)
-        {
-            return await UserDialogs.Instance.ConfirmAsync(config);
-        }
+        
     }
 }
